feat: add utils for helper funtions
